setAuthUser function
